# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: KSA Rangeland Intelligence API
  version: 0.1.0
  description: >
    Backend contract for the KSA / WFP disaster-monitoring platform. The frontend
    lets a user pick a location and a disaster indicator; this API returns the
    indicator time-series plus a model insight. The insight is split into two
    machine spaces: deterministic analytics (severity, trend, anomaly) computed by
    code, and a plain-language narrative generated by the local LLM service.

    This spec is the single source of truth. `contracts/types.ts` mirrors it for the
    frontend; `services/api/app/schemas.py` mirrors it for the backend. The
    `test_contract` gate test asserts the served schema stays in sync.

servers:
  - url: /api
    description: Reached through the Nginx reverse proxy on the physical server.

paths:
  /health:
    get:
      operationId: getHealth
      summary: Liveness + data-source status.
      responses:
        "200":
          description: Service healthy.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Health" }

  /locations:
    get:
      operationId: listLocations
      summary: Monitored locations (ASAL counties).
      responses:
        "200":
          description: Location list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Location" }

  /indicators:
    get:
      operationId: listIndicators
      summary: Disaster indicators available for monitoring.
      responses:
        "200":
          description: Indicator list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Indicator" }

  /observations:
    get:
      operationId: getObservations
      summary: Indicator time-series for one location.
      parameters:
        - { name: location,  in: query, required: true,  schema: { type: string }, description: Location id. }
        - { name: indicator, in: query, required: true,  schema: { type: string }, description: Indicator id. }
        - { name: from,      in: query, required: false, schema: { type: string }, description: Inclusive start, YYYY-MM. }
        - { name: to,        in: query, required: false, schema: { type: string }, description: Inclusive end, YYYY-MM. }
      responses:
        "200":
          description: Series.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ObservationSeries" }
        "404":
          description: Unknown location or indicator.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /insight:
    post:
      operationId: getInsight
      summary: Deterministic analytics plus generated narrative for a location + indicator.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InsightRequest" }
      responses:
        "200":
          description: Insight.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Insight" }
        "404":
          description: Unknown location or indicator.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

components:
  schemas:
    DataSourceKind:
      type: string
      description: Provenance of the data. `stub` is clearly-labeled placeholder data used until real feeds are wired; it is never real.
      enum: [stub, geoserver]

    Health:
      type: object
      required: [status, data_source, llm_mode]
      properties:
        status:      { type: string, enum: [ok] }
        data_source: { $ref: "#/components/schemas/DataSourceKind" }
        llm_mode:    { type: string, enum: [claude, stub], description: Whether narrative comes from local Claude Code or a templated stub. }
        version:     { type: string }

    LatLon:
      type: object
      required: [lat, lon]
      properties:
        lat: { type: number }
        lon: { type: number }

    Location:
      type: object
      required: [id, name, type, centroid]
      properties:
        id:       { type: string, example: turkana }
        name:     { type: string, example: Turkana }
        type:     { type: string, enum: [county], default: county }
        centroid: { $ref: "#/components/schemas/LatLon" }
        bbox:
          type: array
          description: "[minLon, minLat, maxLon, maxLat]"
          items: { type: number }
          minItems: 4
          maxItems: 4

    IndicatorCategory:
      type: string
      enum: [vegetation, drought, hydrology, hazard]

    Direction:
      type: string
      description: Whether a higher value is a better or worse condition. Drives severity interpretation.
      enum: [higher_is_better, lower_is_better]

    Indicator:
      type: object
      required: [id, name, full_name, unit, category, direction]
      properties:
        id:          { type: string, example: vci }
        name:        { type: string, example: VCI }
        full_name:   { type: string, example: Vegetation Condition Index }
        unit:        { type: string, example: percent }
        category:    { $ref: "#/components/schemas/IndicatorCategory" }
        direction:   { $ref: "#/components/schemas/Direction" }
        description: { type: string }
        value_range:
          type: object
          properties:
            min: { type: number }
            max: { type: number }

    ObservationPoint:
      type: object
      required: [t, value]
      properties:
        t:     { type: string, description: Period label, YYYY-MM. }
        value: { type: number, nullable: true }

    ObservationSeries:
      type: object
      required: [location, indicator, unit, source, points]
      properties:
        location:  { type: string }
        indicator: { type: string }
        unit:      { type: string }
        source:    { $ref: "#/components/schemas/DataSourceKind" }
        points:
          type: array
          items: { $ref: "#/components/schemas/ObservationPoint" }

    SeverityClass:
      type: string
      description: Aligned with the alert levels the UI already renders.
      enum: [normal, watch, warning, severe, emergency]

    TrendDirection:
      type: string
      enum: [improving, stable, declining]

    Trend:
      type: object
      required: [direction, slope_per_month, pct_change]
      properties:
        direction:       { $ref: "#/components/schemas/TrendDirection" }
        slope_per_month: { type: number }
        pct_change:      { type: number, description: Percent change over the window, first vs last. }

    Anomaly:
      type: object
      required: [z_score, is_anomalous]
      properties:
        z_score:      { type: number }
        is_anomalous: { type: boolean, description: "True when |z| >= 2." }

    Severity:
      type: object
      required: [class, score, rationale_code]
      properties:
        class:          { $ref: "#/components/schemas/SeverityClass" }
        score:          { type: number, description: 0 (best) to 10 (worst). }
        rationale_code: { type: string, description: Machine-readable reason, e.g. vci_below_35. }

    Analytics:
      type: object
      required: [latest_value, mean, trend, anomaly, severity, n_points]
      properties:
        latest_value: { type: number, nullable: true }
        mean:         { type: number, nullable: true }
        n_points:     { type: integer }
        trend:        { $ref: "#/components/schemas/Trend" }
        anomaly:      { $ref: "#/components/schemas/Anomaly" }
        severity:     { $ref: "#/components/schemas/Severity" }

    Narrative:
      type: object
      required: [summary, recommendation, generated, model]
      properties:
        summary:        { type: string }
        recommendation: { type: string }
        generated:      { type: boolean, description: True when produced by local Claude Code, false when templated stub. }
        model:          { type: string }

    InsightRequest:
      type: object
      required: [location, indicator]
      properties:
        location:  { type: string }
        indicator: { type: string }
        from:      { type: string, description: Inclusive start, YYYY-MM. }
        to:        { type: string, description: Inclusive end, YYYY-MM. }

    Period:
      type: object
      properties:
        from: { type: string }
        to:   { type: string }

    Insight:
      type: object
      required: [location, indicator, period, series, analytics, narrative, data_source]
      properties:
        location:    { $ref: "#/components/schemas/Location" }
        indicator:   { $ref: "#/components/schemas/Indicator" }
        period:      { $ref: "#/components/schemas/Period" }
        series:      { $ref: "#/components/schemas/ObservationSeries" }
        analytics:   { $ref: "#/components/schemas/Analytics" }
        narrative:   { $ref: "#/components/schemas/Narrative" }
        data_source: { $ref: "#/components/schemas/DataSourceKind" }
        generated_at: { type: string, nullable: true, description: ISO-8601, stamped by the server. }

    Error:
      type: object
      required: [error, detail]
      properties:
        error:  { type: string }
        detail: { type: string }